Módulo de usuarios 100% completo

+++ src/components/deleteUserComponent/deleteUserComponent.js
@@ -7,6 +7,12 @@ import swal from 'sweetalert'
 import { Redirect } from 'react-router-dom'
 
 export default class deleteUserComponent extends Component {
+    constructor(props) {
+        super(props)
+        this.state = {
+            redirect: false
+        }
+    }
     componentDidMount() {
         // DELETE USER
         let token = JSON.parse(sessionStorage.getItem('token'))
@@ -14,29 +20,67 @@ export default class deleteUserComponent extends Component {
         if (token === null) {
             return <Redirect to={'/'} />
         }
+
         let id = this.props.location.state.userSelected.id
         const config = {
             headers: { Authorization: `Bearer ${token}` }
         };
 
-        axios.delete(Global.getUsers + '/' + id, config)
-            .then(response => {
-                console.log(response)
-                sessionStorage.setItem('token', JSON.stringify(response.data.loggedUser.token))
-            })
-            .catch(e => {
-                if (!e.response.data.Success && e.response.data.HttpCodeResponse === 401) {
-                    HELPER_FUNCTIONS.logout()
+        swal({
+            title: "¿Estás seguro?",
+            text: "Una vez borrado el usuario, no podrás recuperarlo",
+            icon: "warning",
+            buttons: true,
+            dangerMode: true,
+        })
+            .then((willDelete) => {
+                if (willDelete) {
+                    axios.delete(Global.getUsers + '/' + id, config)
+                        .then(response => {
+                            sessionStorage.setItem('token', JSON.stringify(response.data.loggedUser.token))
+                            swal("El usuario ha sido borrado", {
+                                icon: "success",
+                            });
+
+                            // Seteo de estado para el redirect
+                            this.setState({
+                                redirect: true
+                            })
+                        })
+                        .catch(e => {
+                            if (!e.response.data.Success && e.response.data.HttpCodeResponse === 401) {
+                                HELPER_FUNCTIONS.logout()
+                            } else {
+                                sessionStorage.setItem('token', JSON.stringify(e.response.data.loggedUser.token))
+                                swal("Error!", "Hubo un problema", "error");
+
+                                // Seteo de estado para el redirect
+                                this.setState({
+                                    redirect: true
+                                })
+                            }
+                            console.log("Error: ", e)
+                        })
+
                 } else {
-                    sessionStorage.setItem('token', JSON.stringify(e.response.data.loggedUser.token))
-                    swal("Error!", "Hubo un problema", "error");
+                    swal("Casi! El usuario NO ha sido borrado", {
+                        icon: "info",
+                    });
+
+                    // Seteo de estado para el redirect
+                    this.setState({
+                        redirect: true
+                    })
+
                 }
-                console.log("Error: ", e)
-            })
-        // FIN DELETE USER
+            });
     }
 
     render() {
+        // Si se cumple algún request, redirijo a /users
+        if (this.state.redirect) {
+            return <Redirect to="/users" />
+        }
         return (
             <div>
                 <div className="header">
